requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
openai-whisper>=20240314
numpy>=1.24.0
requests>=2.31.0
gunicorn>=21.2.0
